more work on input thread
